Bugfix
Replaced one bug with another. (Programmer of the year award goes to me)
While the current error is now "CRITICAL" instead of "WARN" the program itself is more stable and probably faster?

Added a dispatcher to properly handle multi threaded stuff.
Added a mutex and a queue (Probably not needed)