Cleanup
Minor cleanup.
No need to print errors honestly, This is a graphical application pretty sure if something goes wrong you would see it.

The better way of handling this would be to add an error label in the program itself to alert the user if something has gone wrong.

Can't be bothered to add it tho!!